레시피 기능 추가

+++ src/main/java/Swyp8/Team12/domain/ingredient/entity/Ingredient.java
@@ -23,7 +23,7 @@ public class Ingredient {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ing_id")
-    private int ingId;
+    private Long ingId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
